+++ crates/testing/tests/tests_1/da_task.rs
@@ -15,7 +15,7 @@ use hotshot_testing::{
 };
 use hotshot_types::{
     data::{null_block, ViewNumber},
-    events::HotShotEvent::*,
+    hotshot_event::HotShotEvent::*,
     simple_vote::DAData,
     traits::{
         block_contents::precompute_vid_commitment, election::Membership,
